backend/requirements.txt
django
Pillow
psycopg2-binary
redis
celery
flower
python-dotenv